chore: refactor component props style computation as an interpolation
