+++ packages/system/src/system.tsx
@@ -3,6 +3,7 @@ import createStyled from "./create-styled"
 import { As, ChakraComponent, Options } from "./system.types"
 import {
   applyProp,
+  componentProps,
   domElements,
   DOMElements,
   pseudoProps,
@@ -12,6 +13,7 @@ import {
 
 function styled<T extends As, P = {}>(component: T, options?: Options<T, P>) {
   return createStyled<T, P>(component, options)(
+    componentProps(options),
     systemProps,
     pseudoProps,
     applyProp(component),
